fixes #21
I fixed the issue, but I have no idea what you want to do with the image... I mean, there is a big difference between the proportion of the shape of the intervention agents and the gif images. So I implemented to mode: fit_to_shape or not.
If you set fit_to_shape to true, the image will be distorded to fit the image, in the case where this bool is set to false, the image will conserved its proportion.